# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * PlayEarthPark Lost and Found Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}