# lib/api-zod/src/generated/types/lostReportInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * PlayEarthPark Lost and Found Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { LostReportInputCategory } from './lostReportInputCategory';

export interface LostReportInput {
  ownerName: string;
  ownerContact: string;
  lostDatetimeFrom?: Date;
  lostDatetimeTo?: Date;
  lostLocationEstimated?: string;
  category: LostReportInputCategory;
  features?: string;
}