# lib/api-zod/src/generated/types/lostReportList.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * PlayEarthPark Lost and Found Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { LostReport } from './lostReport';

export interface LostReportList {
  items: LostReport[];
  total: number;
  page: number;
  limit: number;
}