# lib/api-zod/src/generated/types/matchCandidate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * PlayEarthPark Lost and Found Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { FoundItem } from './foundItem';
import type { LostReport } from './lostReport';

export interface MatchCandidate {
  id: string;
  foundItem: FoundItem;
  lostReport: LostReport;
  /** Match score 0-100 */
  score: number;
  /** Human-readable reasons for the match */
  reasons: string[];
}